Remove even more argument type constraints and deprecate XYZ
I have made a mess of argument types, and further testing revealed
that the type constraints combined with the default height vector
filled with integer zeros lead to a MethodError when using floats
for the latitude or longitude values.

Also, the WGS-84 XYZ input method is ugly and should be changed to
accept a matrix or an array of tuples rather than a tuple of arrays.